# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Marketing Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients/{clientId}/posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List posts for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, approved, scheduled, published, failed]
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post (save draft)
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /clients/{clientId}/posts/{postId}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostBody"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /clients/{clientId}/posts/{postId}/publish:
    post:
      operationId: publishPost
      tags: [posts]
      summary: Publish an approved post to its target social platform
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Published post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "422":
          description: No connected account for platform
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clients/{clientId}/posts/{postId}/approve:
    post:
      operationId: approvePost
      tags: [posts]
      summary: Approve a draft post
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovePostBody"
      responses:
        "200":
          description: Approved post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /clients/{clientId}/posts/{postId}/reject:
    post:
      operationId: rejectPost
      tags: [posts]
      summary: Reject a draft post
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rejected post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /clients/{clientId}/posts/bulk-approve:
    post:
      operationId: bulkApprovePosts
      tags: [posts]
      summary: Bulk approve multiple draft posts
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkApproveBody"
      responses:
        "200":
          description: Bulk approval result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkApproveResponse"
...
  /clients/{clientId}/posts/export:
    get:
      operationId: exportPosts
      tags: [posts]
      summary: Export approved posts as Postiz-compatible JSON
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Export data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostizExport"

  # ── Images ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Client ──────────────────────────────────────────────────────────────
...
    Post:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        campaignId:
          type: string
        topic:
          type: string
        caption:
          type: string
        hashtags:
          type: string
        selectedImageUrl:
          type: string
        status:
          type: string
          enum: [draft, approved, scheduled, published, failed, rejected]
        scheduledAt:
          type: string
        platform:
          type: string
          enum: [instagram, facebook, twitter, linkedin, blog, newsletter]
        postType:
          type: string
          enum: [social, blog, newsletter]
        title:
          type: string
        longFormBody:
          type: string
        generationStatus:
          type: string
          enum: [generating, ready, failed]
        imagePrompt:
          type: string
        publishedAt:
          type: string
        publishedUrl:
          type: string
        publishError:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, clientId, topic, caption, status, postType, createdAt, updatedAt]
    CreatePostBody:
      type: object
      properties:
        topic:
          type: string
        caption:
          type: string
        hashtags:
          type: string
        selectedImageUrl:
          type: string
        platform:
          type: string
          enum: [instagram, facebook, twitter, linkedin, blog, newsletter]
        campaignId:
          type: string
        postType:
          type: string
          enum: [social, blog, newsletter]
        title:
          type: string
        longFormBody:
          type: string
      required: [topic, caption]
    UpdatePostBody:
      type: object
      properties:
        topic:
          type: string
        caption:
          type: string
        hashtags:
          type: string
        selectedImageUrl:
          type: string
        status:
          type: string
          enum: [draft, approved, scheduled, published, failed, rejected]
        scheduledAt:
          type: string
        platform:
          type: string
          enum: [instagram, facebook, twitter, linkedin, blog, newsletter]
        campaignId:
          type: string
        postType:
          type: string
          enum: [social, blog, newsletter]
        title:
          type: string
        longFormBody:
          type: string
    BulkApproveBody:
      type: object
      properties:
        postIds:
          type: array
          items:
            type: string
        scheduledAt:
          type: string
        platform:
          type: string
          enum: [instagram, facebook, twitter, linkedin, blog, newsletter]
      required: [postIds]
    BulkApproveResponse:
      type: object
      properties:
        approved:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        count:
          type: number
      required: [approved, count]
...
    ApprovePostBody:
      type: object
      properties:
        scheduledAt:
          type: string
        platform:
          type: string
          enum: [instagram, facebook, twitter, linkedin, blog]
      required: [scheduledAt]
    PostizExport:
      type: object
      properties:
        clientName:
          type: string
        exportedAt:
          type: string
        posts:
          type: array
          items:
            $ref: "#/components/schemas/PostizPost"
      required: [clientName, exportedAt, posts]
    PostizPost:
      type: object
      properties:
        id:
          type: string
        caption:
          type: string
        hashtags:
          type: string
        image_url:
          type: string
        scheduled_at:
          type: string
        platform:
          type: string
      required: [id, caption, scheduled_at]

    # ── Image ─────────────────────────────────────────────────────────────────
...